+++ src/ui/help/mod.rs
@@ -1,0 +1,2 @@
+pub(crate) mod help_widget;
+pub(crate) mod settings_panel;
